chore: fix narrowing conversion in isWordStart

+++ src/common.h
@@ -81,7 +81,7 @@ extern Element ToLower(const Element &s);
 extern Element ToUpper(const Element &s);
 
 extern bool isMatch(const CandidateString &subject, const Element &query_lw, const Element &query_up);
-extern bool isWordStart(const int pos, const CandidateString &subject, const CandidateString &subject_lw);
+extern bool isWordStart(const size_t pos, const CandidateString &subject, const CandidateString &subject_lw);
 extern Score scoreCharacter(const int i, const int j, const bool start, const Score acro_score, const Score csc_score);
 extern Score scoreConsecutives(const CandidateString &subject, const CandidateString &subject_lw, const Element &query, const Element &query_lw, unsigned i, unsigned j, const bool startOfWord);
 extern AcronymResult scoreAcronyms(const CandidateString &subject, const CandidateString &subject_lw, const Element &query, const Element &query_lw);
